fixing circular deps with asset-swapper/types and utils/rfq-client. moved some things from asset-swapper/types to top level types.ts
